Added weight-gen cli
