+++ bencher/Cargo.toml
@@ -7,10 +7,15 @@ version = "0.1.0"
 authors = ["Laminar Developers <hello@laminar.one>"]
 edition = "2018"
 
+[[bin]]
+path = "src/bin.rs"
+name = "codegen"
+
 [dependencies]
 linregress = { version = "0.4.0", optional = true }
 handlebars = {version = "3.5.2", optional = true }
 serde = { features = ['derive'], optional = true, version = "1.0.119" }
+serde_json = "1.0"
 codec = { package = "parity-scale-codec", version = "2.0.0", features = ["derive"], default-features = false }
 sp-core = { version = "3.0.0", default-features = false }
 sp-std = { version = "3.0.0", default-features = false }
