+++ weight-gen/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "weight-gen"
+description = "CLI for generating weight from bencher output"
+license = "Apache-2.0"
+version = "0.1.0"
+authors = ["Laminar Developers <hello@laminar.one>"]
+edition = "2018"
+
+[dependencies]
+serde = { features = ['derive'], optional = true, version = "1.0.119" }
+serde_json = "1.0"
+clap = "3.0.0-beta.2"
+handlebars = {version = "3.5.2", optional = true }
+
+[features]
+default = ["std"]
+std = [
+    "handlebars",
+    "serde/std",
+]
